Upload File request with New UI
Two things to note:
handleDone in FileUpload resets "files" in State Hook, after upload. Otherwise it retained the files in Modal. When we click add again, we don't want the stuff from previous upload.

Second:
The prop passed to Tablefiles from App.js is "", so it doesn't display the dummy data on the table.

Currently the name of the file uploaded has yet to appear on the table, but I can confirm, it's saved in the replica node and received by the central node.

The next step is to pass the name of the files in the system (from a getlist post to the server) to the Tablefiles component.  This should be done, ideally right after the uploadfile post request is sent, so as to update the table with the most current information.